+++ src/main/java/com/mercadolibre/dto/tracking/TrackingData.java
@@ -12,4 +12,5 @@ public class TrackingData {
   private String amount;
   private String installments;
   private String frictionless;
+  private String cardSize;
 }